backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
requests
pypdf
python-docx
sentence-transformers
faiss-cpu
numpy
